moved scripts into copro folder for executable
