+++ setup.py
@@ -33,14 +33,18 @@ setup(
         'Programming Language :: Python :: 3.8',
     ],
     description="modelling interplay between conflict and climate",
-    entry_points={},
+    entry_points={
+        'console_scripts': [
+            'copro_runner=copro.scripts.copro_runner:cli',
+        ],
+    },
     install_requires=requirements,
     license="MIT",
-    long_description='blablabla',
+    long_description='Python-model build on scikit-learn functions, designed to facilitate the set-up, execution, and evaluation of machine-learning models for the study of the climate-conflict nexus',
     include_package_data=True,
     keywords='conflict, climate, machine learning, projections',
     name='copro',
-    packages=find_packages(include=['copro', 'copro.*']),
+    packages=find_packages(exclude=['docs', 'tests', 'joss']),
     setup_requires=setup_requirements,
     test_suite='tests',
     tests_require=test_requirements,
